moving event handler logic to domain model
