+++ samples/spring-eventsourced-shopping-cart/src/main/java/com/example/shoppingcart/ShoppingCartService.java
@@ -10,11 +10,6 @@ import kalix.springsdk.annotations.EventHandler;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 // tag::class[]
 @Entity(entityKey = "cartId", entityType = "shopping-cart")
@@ -48,10 +43,8 @@ public class ShoppingCartService extends EventSourcedEntity<ShoppingCart> {
 
   @PostMapping("/items/{productId}/remove")
   public Effect<String> removeItem(@PathVariable String productId) {
-    if (findItemByProductId(currentState(), productId).isEmpty()) {
-      return effects()
-          .error(
-              "Cannot remove item " + productId + " because it is not in the cart.");
+    if (currentState().findItemByProductId(productId).isEmpty()) {
+      return effects().error("Cannot remove item " + productId + " because it is not in the cart.");
     }
 
     var event = new ShoppingCartEvent.ItemRemoved(productId);
@@ -71,46 +64,14 @@ public class ShoppingCartService extends EventSourcedEntity<ShoppingCart> {
   // tag::itemAdded[]
   @EventHandler
   public ShoppingCart itemAdded(ShoppingCartEvent.ItemAdded itemAdded) {
-    var item = itemAdded.item();
-    var lineItem = updateItem(item, currentState());
-    List<LineItem> lineItems =
-        removeItemByProductId(currentState(), item.productId());
-    lineItems.add(lineItem);
-    lineItems.sort(Comparator.comparing(LineItem::productId));
-    return currentState().withItems(lineItems);
+    return currentState().onItemAdded(itemAdded);
   }
 
   // end::itemAdded[]
 
   @EventHandler
   public ShoppingCart itemRemoved(ShoppingCartEvent.ItemRemoved itemRemoved) {
-    List<LineItem> items =
-        removeItemByProductId(currentState(), itemRemoved.productId());
-    items.sort(Comparator.comparing(LineItem::productId));
-    return currentState().withItems(items);
+    return currentState().onItemRemoved(itemRemoved);
   }
-
-  // tag::itemAdded[]
-  private LineItem updateItem(
-      LineItem item, ShoppingCart cart) {
-    return findItemByProductId(cart, item.productId())
-        .map(li -> li.withQuantity(li.quantity() + item.quantity()))
-        .orElse(item);
-  }
-
-  private Optional<LineItem> findItemByProductId(
-      ShoppingCart cart, String productId) {
-    Predicate<LineItem> lineItemExists =
-        lineItem -> lineItem.productId().equals(productId);
-    return cart.items().stream().filter(lineItemExists).findFirst();
-  }
-
-  private List<LineItem> removeItemByProductId(
-      ShoppingCart cart, String productId) {
-    return cart.items().stream()
-        .filter(lineItem -> !lineItem.productId().equals(productId))
-        .collect(Collectors.toList());
-  }
-  // end::itemAdded[]
 
 }
