+++ samples/spring-eventsourced-shopping-cart/src/main/java/com/example/shoppingcart/domain/ShoppingCart.java
@@ -1,6 +1,10 @@
 package com.example.shoppingcart.domain;
 
+import java.util.Comparator;
 import java.util.List;
+import java.util.Optional;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 public record ShoppingCart(String cartId, List<LineItem> items) {
 
@@ -10,7 +14,44 @@ public record ShoppingCart(String cartId, List<LineItem> items) {
     }
   }
 
-  public ShoppingCart withItems(List<LineItem> items) {
-    return new ShoppingCart(cartId, items);
+  // tag::itemAdded[]
+  public ShoppingCart onItemAdded(ShoppingCartEvent.ItemAdded itemAdded) {
+    var item = itemAdded.item();
+    var lineItem = updateItem(item, this);
+    List<LineItem> lineItems =
+        removeItemByProductId(this, item.productId());
+    lineItems.add(lineItem);
+    lineItems.sort(Comparator.comparing(LineItem::productId));
+    return new ShoppingCart(cartId, lineItems);
   }
+  // end::itemAdded[]
+
+  public ShoppingCart onItemRemoved(ShoppingCartEvent.ItemRemoved itemRemoved) {
+    List<LineItem> updatedItems =
+        removeItemByProductId(this, itemRemoved.productId());
+    updatedItems.sort(Comparator.comparing(LineItem::productId));
+    return new ShoppingCart(cartId, updatedItems);
+  }
+
+  // tag::itemAdded[]
+
+  private static LineItem updateItem(LineItem item, ShoppingCart cart) {
+    return cart.findItemByProductId(item.productId())
+        .map(li -> li.withQuantity(li.quantity() + item.quantity()))
+        .orElse(item);
+  }
+
+  public Optional<LineItem> findItemByProductId(String productId) {
+    Predicate<LineItem> lineItemExists =
+        lineItem -> lineItem.productId().equals(productId);
+    return items.stream().filter(lineItemExists).findFirst();
+  }
+
+  private static List<LineItem> removeItemByProductId(
+      ShoppingCart cart, String productId) {
+    return cart.items().stream()
+        .filter(lineItem -> !lineItem.productId().equals(productId))
+        .collect(Collectors.toList());
+  }
+  // end::itemAdded[]
 }
